Try uint64_t

+++ src/passes/analysis/coreirjson.cpp
@@ -95,7 +95,7 @@ string Value2Json(Value* v) {
       ret.add(to_string(ci->get()));
     }
     else if (auto cbv = dyn_cast<ConstBitVector>(con)) {
-      ret.add(to_string(cbv->get().to_type<uint>()));
+      ret.add(to_string(cbv->get().to_type<uint64_t>()));
     }
     else if (auto cs = dyn_cast<ConstString>(con)) {
       ret.add(quote(cs->get()));
